Verify and correctly handle null/undefined in max/min constraints
Numerically undefined/null should always be false, which is the case.  This just proves it.
Lengthwise, undefined/null is never longer than any given minimum, and is aways shorter than any given maximum.  If a user needs to limit the length to a maximum AND make sure that the object is not undefined or null, then the user should add the isNotNull constraint to the set of constriants.